Move function UiHelper.getFirstVisibleVerseAnchor to verseListController
